[test] Mark test_only_one_lazy_module_initialization as expected failure
When all tests are run together, the "sqlite" adapter may have already
been initialized by another test. So
test_only_one_lazy_module_initialization will break on its first
assertion. Following review discussion, there appears to be no simple
way to handle this so we mark the test as expected failure so that it
does not break in CI environment (or tox) but it might still succeed
when unittest_sqlite.py is run independently.