Fix SQL transformation of COUNTDISTINCT()
This follows-up on changeset 9f2071ae3b26 (2013) in which this
COUNTDISTINCT function got introduced to replace COUNT(X, True) usage
(deprecated since version 1.10). Unfortunately, the implementation is
buggy as it uses the "name" attribute of the COUNTDISTINCT class to
generate the equivalent SQL and there exists no COUNTDISTINCT keyword in
SQL. I think the intended implementation is to convert COUNTDISTINCT(X)
into COUNT(DISTINCT X) instead (as this is done in the
backward-compatibility layer in COUNT.as_sql()).

There is no test for all this unfortunately.