Add support for autopkgtest
In order to get the tests in the source Debian package we need
to ship them in the Python package. That's why we turn the
'test' directory into a Python package and remove the explicit
ignore from `setup.py`.

The tests themselves will install PostgreSQL and MariaDB. They
are thus marked `isolation-container` and need at least the
the lxc autopkgtest runner to be started.